MATLAB Code primary upload
This is all that I've written in MATLAB thus far for this project. I may end up re-organizing it, and implementing some of the files as sub-functions of a single file for example. But for now I just want to get it all in the repository. There are also some images representing some pipe cross sections of interest.